+++ integrations/github/github/helpers/utils.py
@@ -9,3 +9,4 @@ class GithubClientType(StrEnum):
 class ObjectKind(StrEnum):
     REPOSITORY = "repository"
     WORKFLOW = "workflow"
+    WORKFLOW_RUN = "workflow-run"