+++ integrations/github/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Features
 
 - add support for Github workflows
+- add support for Github workflow runs
 
 ## 0.1.0-dev (2025-05-12)